+++ pyanalyze/implementation.py
@@ -1,6 +1,6 @@
 from .annotations import type_from_value
 from .error_code import ErrorCode
-from .extensions import reveal_type
+from .extensions import assert_type, reveal_locals, reveal_type
 from .format_strings import parse_format_string
 from .predicates import IsAssignablePredicate
 from .safe import safe_hasattr, safe_isinstance, safe_issubclass
@@ -50,6 +50,7 @@ from .value import (
     concrete_values_from_iterable,
     kv_pairs_from_mapping,
     make_weak,
+    unannotate,
     unite_values,
     flatten_values,
     replace_known_sequence_value,
@@ -64,6 +65,8 @@ import qcore
 import inspect
 import warnings
 from types import FunctionType
+import typing
+import typing_extensions
 from typing import (
     Sequence,
     TypeVar,
@@ -126,12 +129,23 @@ def _issubclass_impl(ctx: CallContext) -> Value:
     return annotate_with_constraint(TypedValue(bool), constraint)
 
 
-def _isinstance_impl(ctx: CallContext) -> ImplReturn:
+def _isinstance_impl(ctx: CallContext) -> Value:
     class_or_tuple = ctx.vars["class_or_tuple"]
     varname = ctx.varname_for_arg("obj")
-    return ImplReturn(
-        TypedValue(bool), _constraint_from_isinstance(varname, class_or_tuple)
-    )
+    if varname is None or not isinstance(class_or_tuple, KnownValue):
+        return TypedValue(bool)
+    if isinstance(class_or_tuple.val, type):
+        narrowed_type = TypedValue(class_or_tuple.val)
+    elif isinstance(class_or_tuple.val, tuple) and all(
+        isinstance(elt, type) for elt in class_or_tuple.val
+    ):
+        vals = [TypedValue(elt) for elt in class_or_tuple.val]
+        narrowed_type = unite_values(*vals)
+    else:
+        return TypedValue(bool)
+    predicate = IsAssignablePredicate(narrowed_type, ctx.visitor, positive_only=False)
+    constraint = Constraint(varname, ConstraintType.predicate, True, predicate)
+    return annotate_with_constraint(TypedValue(bool), constraint)
 
 
 def _constraint_from_isinstance(
@@ -938,23 +952,41 @@ def _assert_is_value_impl(ctx: CallContext) -> Value:
 
 
 def _reveal_type_impl(ctx: CallContext) -> Value:
+    value = ctx.vars["value"]
     if ctx.visitor._is_checking():
-        value = ctx.vars["value"]
         message = f"Revealed type is {ctx.visitor.display_value(value)}"
         ctx.show_error(message, ErrorCode.inference_failure, arg="value")
+    return value
+
+
+def _reveal_locals_impl(ctx: CallContext) -> Value:
+    scope = ctx.visitor.scopes.current_scope()
+    if ctx.visitor._is_collecting():
+        for varname in scope.all_variables():
+            scope.get(varname, ctx.node, ctx.visitor.state)
+    else:
+        details = []
+        for varname in scope.all_variables():
+            val, _, _ = scope.get(varname, ctx.node, ctx.visitor.state)
+            details.append(CanAssignError(f"{varname}: {val}"))
+        ctx.show_error(
+            "Revealed local types are:",
+            ErrorCode.inference_failure,
+            detail=str(CanAssignError(children=details)),
+        )
     return KnownValue(None)
 
 
 def _dump_value_impl(ctx: CallContext) -> Value:
+    value = ctx.vars["value"]
     if ctx.visitor._is_checking():
-        value = ctx.vars["value"]
         message = f"Value is '{value!r}'"
         if isinstance(value, KnownValue):
             sig = ctx.visitor.arg_spec_cache.get_argspec(value.val)
             if sig is not None:
                 message += f", signature is {sig!r}"
         ctx.show_error(message, ErrorCode.inference_failure, arg="value")
-    return KnownValue(None)
+    return value
 
 
 def _str_format_impl(ctx: CallContext) -> Value:
@@ -1040,6 +1072,20 @@ def _cast_impl(ctx: CallContext) -> Value:
     return type_from_value(typ, visitor=ctx.visitor, node=ctx.node)
 
 
+def _assert_type_impl(ctx: CallContext) -> Value:
+    # TODO maybe we should walk over the whole value and remove Annotated.
+    val = unannotate(ctx.vars["val"])
+    typ = ctx.vars["typ"]
+    expected_type = type_from_value(typ, visitor=ctx.visitor, node=ctx.node)
+    if val != expected_type:
+        ctx.show_error(
+            f"Type is {val} (expected {expected_type})",
+            error_code=ErrorCode.inference_failure,
+            arg="obj",
+        )
+    return val
+
+
 def _subclasses_impl(ctx: CallContext) -> Value:
     """Overridden because typeshed types make it (T) => List[T] instead."""
     self_obj = ctx.vars["self"]
@@ -1104,6 +1150,7 @@ _ENCODING_PARAMETER = SigParameter(
     "encoding", annotation=TypedValue(str), default=KnownValue("")
 )
 
+T = TypeVar("T")
 K = TypeVar("K")
 V = TypeVar("V")
 
@@ -1132,14 +1179,17 @@ def get_default_argspecs() -> Dict[object, Signature]:
             callable=assert_is_value,
         ),
         Signature.make(
-            [SigParameter("value", _POS_ONLY)],
-            KnownValue(None),
+            [SigParameter("value", _POS_ONLY, annotation=TypeVarValue(T))],
+            TypeVarValue(T),
             impl=_reveal_type_impl,
             callable=reveal_type,
         ),
         Signature.make(
-            [SigParameter("value", _POS_ONLY)],
-            KnownValue(None),
+            [], KnownValue(None), impl=_reveal_locals_impl, callable=reveal_locals
+        ),
+        Signature.make(
+            [SigParameter("value", _POS_ONLY, annotation=TypeVarValue(T))],
+            TypeVarValue(T),
             impl=_dump_value_impl,
             callable=dump_value,
         ),
@@ -1420,7 +1470,18 @@ def get_default_argspecs() -> Dict[object, Signature]:
             callable=str.format,
         ),
         Signature.make(
-            [SigParameter("typ"), SigParameter("val")], callable=cast, impl=_cast_impl
+            [SigParameter("typ", _POS_ONLY), SigParameter("val", _POS_ONLY)],
+            callable=cast,
+            impl=_cast_impl,
+        ),
+        Signature.make(
+            [
+                SigParameter("val", _POS_ONLY, annotation=TypeVarValue(T)),
+                SigParameter("typ", _POS_ONLY),
+            ],
+            TypeVarValue(T),
+            callable=assert_type,
+            impl=_assert_type_impl,
         ),
         # workaround for https://github.com/python/typeshed/pull/3501
         Signature.make(
@@ -1549,4 +1610,47 @@ def get_default_argspecs() -> Dict[object, Signature]:
             ),
         ),
     ]
+    for mod in typing, typing_extensions:
+        try:
+            reveal_type_func = getattr(mod, "reveal_type")
+        except AttributeError:
+            pass
+        else:
+            # Anticipating https://bugs.python.org/issue46414
+            sig = Signature.make(
+                [SigParameter("value", _POS_ONLY, annotation=TypeVarValue(T))],
+                TypeVarValue(T),
+                impl=_reveal_type_impl,
+                callable=reveal_type_func,
+            )
+            signatures.append(sig)
+        # Anticipating that this will be added to the stdlib
+        try:
+            assert_type_func = getattr(mod, "assert_type")
+        except AttributeError:
+            pass
+        else:
+            sig = Signature.make(
+                [
+                    SigParameter("val", _POS_ONLY, annotation=TypeVarValue(T)),
+                    SigParameter("typ", _POS_ONLY),
+                ],
+                TypeVarValue(T),
+                callable=assert_type_func,
+                impl=_assert_type_impl,
+            )
+            signatures.append(sig)
+        # Anticipating that this will be added to the stdlib
+        try:
+            reveal_locals_func = getattr(mod, "reveal_locals")
+        except AttributeError:
+            pass
+        else:
+            sig = Signature.make(
+                [],
+                KnownValue(None),
+                callable=reveal_locals_func,
+                impl=_reveal_locals_impl,
+            )
+            signatures.append(sig)
     return {sig.callable: sig for sig in signatures}